requirements.txt
amadeus==9.0.0
openai>=1.40.0
pydantic>=2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0